Move regression-specific filtersets to test module
